Renames types file
